Remove IterableOnce.
