+++ src/main/scala/strawman/collection/javaSupport.scala
@@ -56,7 +56,7 @@ class StringOps(val s: String)
   /** Overloaded version of `++` that gives back a string, where the inherited
     *  version gives back a sequence.
     */
-  def ++(xs: IterableOnce[Char]): String = {
+  def ++(xs: Iterable[Char]): String = {
     val sb = new StringBuilder() ++= s
     for (ch <- xs.iterating) sb += ch
     sb.result
@@ -100,9 +100,9 @@ class ArrayOps[A](val xs: Array[A])
   override def className = "Array"
 
   def map[B: ClassTag](f: A => B): Array[B] = fromIterable(View.Map(coll, f))
-  def flatMap[B: ClassTag](f: A => IterableOnce[B]): Array[B] = fromIterable(View.FlatMap(coll, f))
-  def ++[B >: A : ClassTag](xs: IterableOnce[B]): Array[B] = fromIterable(View.Concat(coll, xs))
-  def zip[B: ClassTag](xs: IterableOnce[B]): Array[(A, B)] = fromIterable(View.Zip(coll, xs))
+  def flatMap[B: ClassTag](f: A => Iterable[B]): Array[B] = fromIterable(View.FlatMap(coll, f))
+  def ++[B >: A : ClassTag](xs: Iterable[B]): Array[B] = fromIterable(View.Concat(coll, xs))
+  def zip[B: ClassTag](xs: Iterable[B]): Array[(A, B)] = fromIterable(View.Zip(coll, xs))
 }
 
 case class ArrayView[A](xs: Array[A]) extends IndexedView[A] {
